setup the files such that types is no longer required.
